research/pyproject.toml
[tool.poetry]
name = "research"
version = "0.1.0"
description = ""
authors = ["isheng"]
#readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
ortools = "^9.11.4210"
googlemaps = "^4.10.0"
python-dotenv = "^1.0.1"
langchain-openai = "^0.3.32"
langchain = "^0.3.27"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
